Separate CsvSource from the csv_generation file
