+++ usaspending_api/download/filestreaming/csv_generation.py
@@ -16,53 +16,15 @@ from usaspending_api.awards.v2.lookups.lookups import contract_type_mapping, ass
 from usaspending_api.common.helpers.generic_helper import generate_raw_quoted_query
 from usaspending_api.download.helpers import (verify_requested_columns_available, multipart_upload, split_csv,
                                               write_to_download_log as write_to_log)
+from usaspending_api.download.filestreaming.csv_source import CsvSource
 from usaspending_api.download.lookups import JOB_STATUS_DICT, VALUE_MAPPINGS
-from usaspending_api.download.v2 import download_column_historical_lookups
-from usaspending_api.references.models import ToptierAgency
 
 DOWNLOAD_VISIBILITY_TIMEOUT = 60*10
 MAX_VISIBILITY_TIMEOUT = 60*60*4
-BUFFER_SIZE = (5 * 1024 ** 2)
 EXCEL_ROW_LIMIT = 1000000
 WAIT_FOR_PROCESS_SLEEP = 5
 
 logger = logging.getLogger('console')
-
-
-class CsvSource:
-    def __init__(self, model_type, file_type, source_type, agency_id):
-        self.model_type = model_type
-        self.file_type = file_type
-        self.source_type = source_type
-        self.query_paths = download_column_historical_lookups.query_paths[model_type][file_type]
-        self.human_names = list(self.query_paths.keys())
-        if agency_id == 'all':
-            self.agency_code = 'all'
-        else:
-            self.agency_code = ToptierAgency.objects.filter(toptier_agency_id=agency_id).first().cgac_code
-        self.queryset = None
-
-    def values(self, header):
-        query_paths = [self.query_paths[hn] for hn in header]
-        return self.queryset.values_list(query_paths).iterator()
-
-    def columns(self, requested):
-        """Given a list of column names requested, returns the ones available in the source"""
-        result = self.human_names
-        if requested:
-            result = [header for header in requested if header in self.human_names]
-
-        # remove headers that we don't have a query path for
-        result = [header for header in result if header in self.query_paths]
-
-        return result
-
-    def row_emitter(self, headers_requested):
-        headers = self.columns(headers_requested)
-        # Not yielding headers as the files can be split
-        # yield headers
-        query_paths = [self.query_paths[hn] for hn in headers]
-        return self.queryset.values(*query_paths)
 
 
 def generate_csvs(download_job, sqs_message=None):
